NetworkRailProducer/requirements.txt
stomp.py==8.1.0
pytz==2022.1
confluent-kafka==2.4.0
